requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Flask-MySQL==1.5.2
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.6
Jinja2==3.1.2
